Added new test for UniqueCoordinateArrayFilter class. Small fixes related to signed/unsigned comparison.
git-svn-id: http://svn.osgeo.org/geos/trunk@1416 5242fede-7e19-0410-aef8-94bd7d2200fb

+++ tests/unit/geom/GeometryFactoryTest.cpp
@@ -503,7 +503,7 @@ namespace tut
 	template<>
 	void object::test<13>()
 	{
-		const int size = 5;
+		const size_t size = 5;
 		CoordArrayPtr coords = new geos::geom::CoordinateArraySequence(size);
 		ensure( coords != 0 );
 		ensure_equals( coords->getSize(), size );
